Reorganizing after the junction of  sequence and splits
