+++ .gitignore
@@ -169,3 +169,4 @@ cython_debug/
 test/test_composite.py
 test/test_rnn_model.py
 
+results/
