+++ mlex/models/__init__.py
@@ -1,1 +1,2 @@
 from .models import *
+from .composite import *
