--- mlex/torch_modules/__init__.py
@@ -1,2 +1,0 @@
-from composite import *
-from models import *
