--- results/account.csv
@@ -1,83 +1,0 @@
-,NUMERO_BANCO,NUMERO_AGENCIA,NUMERO_CONTA,TIPO,CPF_CNPJ_TITULAR,I-a,PREDICTED,TIPO,CNAB,NATUREZA_SALDO,DIA_LANCAMENTO,MES_LANCAMENTO,VALOR_TRANSACAO,VALOR_SALDO
-60671,N3Z,492,310971,4,MR3838274417,,0.74054855,4,218,C,15,8,0.46,2021.15
-61946,J5A,492,554700,4,MR3838274417,,0.74196225,4,117,D,27,8,219.88,0.0
-62187,N3Z,492,310971,4,MR3838274417,,0.73345673,4,218,C,29,8,0.41,1785.69
-63540,N3Z,492,310971,4,MR3838274417,,0.7956097,4,218,C,10,9,0.21,931.04
-65058,W5Y,16,671769,1,MR3838274417,,0.7426965,1,213,C,20,9,433.0,16973.85
-65392,O6Y,432,729869,2,MR3838274417,,0.76329213,2,114,C,23,9,10.0,8172.47
-65393,W9N,9379,830443,2,MR3838274417,,0.7969065,2,205,C,23,9,0.03,7.35
-65394,N3Z,492,310971,4,MR3838274417,,0.7914004,4,218,C,23,9,0.27,1339.66
-65581,N3Z,492,310971,4,MR3838274417,,0.73880506,4,218,C,24,9,0.28,1339.94
-66000,W5Y,16,671769,1,MR3838274417,,0.743094,1,213,C,27,9,469.0,15213.85
-66902,N3Z,492,310971,4,MR3838274417,,0.7500406,4,218,C,4,10,0.08,388.45
-67154,W9N,9379,830443,1,MR3838274417,,0.7601141,1,112,D,7,10,7250.0,7240.0
-67155,W9N,9379,830443,1,MR3838274417,,0.739799,1,206,C,7,10,7249.99,9.99
-67156,W9N,9379,830443,1,MR3838274417,,0.7882118,1,206,C,7,10,0.01,10.0
-67157,W9N,9379,439917,2,MR3838274417,,0.8219704,2,205,C,7,10,61.94,52102.83
-67158,W9N,9379,439917,2,MR3838274417,,0.7544107,2,205,C,7,10,18.27,52121.1
-70703,W5Y,16,671769,1,MR3838274417,,0.7429469,1,213,C,31,10,435.0,3962.26
-70704,W5Y,16,671769,1,MR3838274417,,0.7419783,1,213,C,31,10,2521.0,6483.26
-70706,O6Y,432,968357,1,MR3838274417,,0.7391916,1,213,C,31,10,464.0,2839.1
-77508,O6Y,432,729869,2,MR3838274417,,0.74077594,2,209,C,18,12,3750.0,11578.12
-77509,O6Y,432,729869,2,MR3838274417,,0.7697671,2,114,C,18,12,10000.0,1578.12
-77510,W9N,9379,830443,1,MR3838274417,,0.78735006,1,112,D,18,12,792.0,782.0
-77511,W9N,9379,830443,1,MR3838274417,,0.7848295,1,117,D,18,12,3500.0,4282.0
-77512,W9N,9379,830443,1,MR3838274417,,0.74942976,1,206,C,18,12,4291.58,9.58
-77513,W9N,9379,830443,1,MR3838274417,,0.745338,1,206,C,18,12,0.42,10.0
-77514,W9N,9379,439917,2,MR3838274417,,0.78578335,2,213,C,18,12,3500.0,63159.31
-77515,N3Z,492,310971,4,MR3838274417,,0.80653715,4,218,C,18,12,0.05,246.31
-77642,N3Z,492,310971,4,MR3838274417,,0.74366325,4,218,C,19,12,0.03,246.34
-99455,N3Z,492,310971,4,MR3838274417,,0.7489156,4,218,C,27,7,0.02,294.21
-105613,E7F,492,411108,1,MR3838274417,,0.74402195,1,120,C,21,8,50.0,13714.36
-107554,W5Y,16,671769,1,MR3838274417,,0.7346277,1,104,C,31,8,164.03,14623.18
-107555,W5Y,16,671769,1,MR3838274417,,0.7309984,1,104,C,31,8,59.99,14563.19
-107558,O6Y,432,729869,2,MR3838274417,,0.7355794,2,209,C,31,8,3573.0,30343.48
-107559,W9N,9379,830443,1,MR3838274417,,0.7406453,1,112,D,31,8,5299.0,5121.0
-109104,O6Y,432,729869,2,MR3838274417,,0.73081416,2,114,C,8,9,518.4,27696.78
-109105,O6Y,432,729869,2,MR3838274417,,0.76288736,2,114,C,8,9,18.0,27678.78
-109106,O6Y,432,729869,2,MR3838274417,,0.78386515,2,114,C,8,9,79.58,27599.2
-109107,O6Y,432,729869,2,MR3838274417,,0.7933504,2,114,C,8,9,470.0,27129.2
-109108,O6Y,432,729869,2,MR3838274417,,0.8448788,2,114,C,8,9,270.0,26859.2
-112415,W9N,9379,830443,2,MR3838274417,,0.736501,2,205,C,22,9,0.01,7.55
-113029,O6Y,432,729869,2,MR3838274417,,0.78301644,2,117,C,24,9,3000.0,20594.8
-113030,O6Y,432,729869,2,MR3838274417,,0.77003145,2,114,C,24,9,3612.79,16982.01
-113729,O6Y,432,729869,2,MR3838274417,,0.8004343,2,114,C,28,9,37.9,24145.88
-113730,O6Y,432,729869,2,MR3838274417,,0.8325122,2,114,C,28,9,74.1,24071.78
-113731,O6Y,432,729869,2,MR3838274417,,0.85421914,2,114,C,28,9,164.67,23907.11
-113732,O6Y,432,729869,2,MR3838274417,,0.86443686,2,105,C,28,9,2.3,23904.81
-113733,O6Y,432,729869,2,MR3838274417,,0.84745896,2,105,C,28,9,2.3,23902.51
-113734,O6Y,432,729869,2,MR3838274417,,0.817988,2,105,C,28,9,2.3,23900.21
-113735,O6Y,432,729869,2,MR3838274417,,0.7823877,2,105,C,28,9,2.3,23897.91
-113736,O6Y,432,729869,2,MR3838274417,,0.7619099,2,105,C,28,9,2.3,23895.61
-113737,O6Y,432,729869,2,MR3838274417,,0.7316565,2,105,C,28,9,2.3,23893.31
-113739,O6Y,432,729869,2,MR3838274417,,0.73856294,2,105,C,28,9,1.1,23891.11
-113740,O6Y,432,729869,2,MR3838274417,,0.7438567,2,105,C,28,9,1.1,23890.01
-113741,O6Y,432,729869,2,MR3838274417,,0.75143045,2,105,C,28,9,1.1,23888.91
-117151,O6Y,432,968357,1,MR3838274417,,0.74608284,1,112,C,13,10,29.5,9816.33
-117152,O6Y,432,968357,1,MR3838274417,,0.80788875,1,112,C,13,10,8.0,9808.33
-117153,O6Y,432,968357,1,MR3838274417,,0.8510059,1,112,C,13,10,119.96,9688.37
-117154,O6Y,432,729869,2,MR3838274417,,0.8768104,2,114,C,13,10,2000.0,5218.55
-117155,O6Y,432,729869,2,MR3838274417,,0.8952789,2,117,C,13,10,3000.0,2218.55
-117156,O6Y,432,729869,2,MR3838274417,,0.89286894,2,117,C,13,10,2000.0,218.55
-117157,O6Y,432,729869,2,MR3838274417,,0.8695774,2,105,C,13,10,1.1,217.45
-117158,O6Y,432,729869,2,MR3838274417,,0.8310331,2,105,C,13,10,1.1,216.35
-117159,O6Y,432,729869,2,MR3838274417,,0.7905494,2,105,C,13,10,1.1,215.25
-117160,W9N,9379,830443,1,MR3838274417,,0.7590664,1,220,C,13,10,866.0,585.51
-120378,O6Y,432,968357,1,MR3838274417,,0.74221903,1,201,C,26,10,832.0,13195.47
-121450,W5Y,16,671769,1,MR3838274417,,0.7908169,1,114,C,29,10,2000.0,2639.23
-121451,O6Y,432,968357,1,MR3838274417,,0.83802545,1,114,C,29,10,2000.0,12046.37
-121452,O6Y,432,729869,2,MR3838274417,,0.8666308,2,114,C,29,10,2000.0,2343.41
-121453,O6Y,432,729869,2,MR3838274417,,0.89086914,2,114,C,29,10,306.2,2037.21
-121454,O6Y,432,729869,2,MR3838274417,,0.9022146,2,105,C,29,10,2.3,2034.91
-122918,W9N,9379,830443,1,MR3838274417,,0.7350671,1,112,D,5,11,167.22,157.22
-122922,W9N,9379,439917,2,MR3838274417,,0.7804627,2,205,C,5,11,2.32,2420.3
-124718,O6Y,432,729869,2,MR3838274417,,0.7563984,2,114,C,13,11,22.5,10026.36
-124719,O6Y,432,729869,2,MR3838274417,,0.8271034,2,114,C,13,11,6.0,10020.36
-124720,W9N,9379,439917,2,MR3838274417,,0.82872117,2,205,C,13,11,0.48,2420.78
-124721,N3Z,492,310971,4,MR3838274417,,0.8112765,4,218,C,13,11,0.02,300.48
-128687,W9N,9379,830443,1,MR3838274417,,0.74803805,1,213,C,4,12,1424.0,1434.0
-128688,W9N,9379,830443,1,MR3838274417,,0.7343975,1,106,C,4,12,1424.0,10.0
-131198,O6Y,432,968357,1,MR3838274417,,0.74924266,1,209,C,18,12,2200.0,8114.92
-131210,N3Z,492,310971,4,MR3838274417,,0.7907897,4,218,C,18,12,0.07,919.91
-132164,O6Y,432,729869,2,MR3838274417,,0.75643516,2,105,C,28,12,2.3,5288.66
-132165,N3Z,492,310971,4,MR3838274417,,0.83303404,4,218,C,28,12,0.08,1000.26
