--- test/__init__.py
@@ -1,2 +1,0 @@
-from .test_simplernn import *
-from .test_rnn_model import *
